# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UK Football Fixed Odds Pool Betting API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginBody" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
        "401":
          description: Invalid credentials
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
        "401":
          description: Not authenticated
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }

  # ---------- PUBLIC ----------
...
  /public/agent-applications:
    post:
      operationId: submitAgentApplication
      tags: [public]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentApplicationBody" }
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentApplication" }

  # ---------- ADMIN ----------
...
  /agent/cashiers:
    get:
      operationId: listCashiers
      tags: [agent]
      responses:
        "200":
          description: Cashiers under this agent
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Cashier" }
    post:
      operationId: createCashier
      tags: [agent]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCashierBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cashier" }

  # ---------- CASHIER ----------
...
  /cashier/weeks/{id}:
    get:
      operationId: getCashierWeekDetail
      tags: [cashier]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Week with fixture statuses (no team names) and odds
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CashierWeekDetail" }
  /cashier/calculate:
    post:
      operationId: calculateBet
      tags: [cashier]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CalculateBetBody" }
      responses:
        "200":
          description: Computed lines and per-line stake
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CalculateBetResult" }

  # ---------- TICKETS ----------
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]

    # ----- Auth -----
    LoginBody:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
      required: [username, password]
    AuthUser:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        name: { type: string }
        role: { type: string, enum: [admin, agent, cashier] }
        agentId: { type: ["integer", "null"] }
        cashierId: { type: ["integer", "null"] }
        shopName: { type: ["string", "null"] }
      required: [id, username, name, role, agentId, cashierId, shopName]

    # ----- Public -----
...
    AgentApplicationBody:
      type: object
      properties:
        shopName: { type: string }
        location: { type: string }
        phone: { type: string }
        email: { type: string }
      required: [shopName, location, phone, email]
    AgentApplication:
      type: object
      properties:
        id: { type: integer }
        shopName: { type: string }
        location: { type: string }
        phone: { type: string }
        email: { type: string }
        status: { type: string, enum: [pending, approved, rejected] }
        createdAt: { type: string, format: date-time }
      required: [id, shopName, location, phone, email, status, createdAt]
...
    Cashier:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        agentId: { type: integer }
        username: { type: string }
        name: { type: string }
        maxStake: { type: number }
        ticketCount: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, userId, agentId, username, name, maxStake, ticketCount, createdAt]
    CreateCashierBody:
     type: object
     properties:
       password: { type: string }
       name: { type: string }
       maxStake: { type: number }
     required: [password, name, maxStake]
...
    WeekOdds:
      type: object
      properties:
        id: { type: integer }
        weekId: { type: integer }
        poolType: { type: string, enum: [single, double, nap, under3, under4, under5, under6] }
        oddsType: { type: string, enum: [standard, high] }
        oddsValue: { type: number }
      required: [id, weekId, poolType, oddsType, oddsValue]
...
    CashierWeek:
      type: object
      properties:
        id: { type: integer }
        weekNumber: { type: integer }
        status: { type: string }
        closesAt: { type: string, format: date-time }
      required: [id, weekNumber, status, closesAt]
    CashierFixture:
      type: object
      properties:
        number: { type: integer }
        status: { type: string, enum: [open, closed, postponed] }
      required: [number, status]
    CashierWeekDetail:
      type: object
      properties:
        week: { $ref: "#/components/schemas/CashierWeek" }
        fixtures:
          type: array
          items: { $ref: "#/components/schemas/CashierFixture" }
        odds:
          type: array
          items: { $ref: "#/components/schemas/WeekOdds" }
        maxStake: { type: number }
      required: [week, fixtures, odds, maxStake]
    CalculateBetBody:
      type: object
      properties:
        weekId: { type: integer }
        betType: { type: string, enum: [nap, perm] }
        poolType: { type: string, enum: [single, double, nap, under3, under4, under5, under6] }
        oddsType: { type: string, enum: [standard, high] }
        selectedNumbers:
          type: array
          items: { type: integer }
        stake: { type: number }
      required: [weekId, betType, poolType, oddsType, selectedNumbers, stake]
    CalculateBetResult:
      type: object
      properties:
        totalLines: { type: integer }
        perLine: { type: number }
        oddsValue: { type: number }
        possibleMaxWinnings: { type: number }
      required: [totalLines, perLine, oddsValue, possibleMaxWinnings]
...