# lib/api-zod/src/generated/types/cashierWeek.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UK Football Fixed Odds Pool Betting API
 * OpenAPI spec version: 0.1.0
 */

export interface CashierWeek {
  id: number;
  weekNumber: number;
  status: string;
  closesAt: Date;
}